+++ src/main/java/blackjack/controller/GameController.java
@@ -3,10 +3,10 @@ package blackjack.controller;
 import blackjack.domain.Statistic;
 import blackjack.domain.card.CardDeck;
 import blackjack.domain.card.CardDeckGenerator;
-import blackjack.domain.human.Dealer;
-import blackjack.domain.human.Name;
-import blackjack.domain.human.Gambler;
-import blackjack.domain.human.Gamblers;
+import blackjack.domain.player.Dealer;
+import blackjack.domain.player.Name;
+import blackjack.domain.player.Gambler;
+import blackjack.domain.player.Gamblers;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
 import java.util.ArrayList;
