+++ src/test/java/blackjack/domain/StatisticTest.java
@@ -5,10 +5,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Denomination;
 import blackjack.domain.card.Symbol;
-import blackjack.domain.human.Dealer;
-import blackjack.domain.human.Name;
-import blackjack.domain.human.Gambler;
-import blackjack.domain.human.Gamblers;
+import blackjack.domain.player.Dealer;
+import blackjack.domain.player.Name;
+import blackjack.domain.player.Gambler;
+import blackjack.domain.player.Gamblers;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
