+++ src/test/java/blackjack/domain/player/GamblersTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.human;
+package blackjack.domain.player;
 
 import static org.assertj.core.api.Assertions.assertThat;
